task 5

+++ 0x0C-more_malloc_free/101-mul.c
@@ -10,7 +10,7 @@
  */
 int is_digit(char *s)
 {
-	int i = 0;
+	long int i = 0;
 
 	while (s[i])
 	{
@@ -29,12 +29,13 @@ int is_digit(char *s)
  */
 int _strlen(char *s)
 {
-	int i = 0;
+	long int i = 0;
 
 	while (s[i] != '\0')
 	{
 		i++;
 	}
+
 	return (i);
 }
 
@@ -57,7 +58,7 @@ void errors(void)
 int main(int argc, char *argv[])
 {
 	char *s1, *s2;
-	int len1, len2, len, i, carry, digit1, digit2, *result, a = 0;
+	long int len1, len2, len, i, carry, digit1, digit2, *result, a = 0;
 
 	s1 = argv[1], s2 = argv[2];
 	if (argc != 3 || !is_digit(s1) || !is_digit(s2))
@@ -65,7 +66,7 @@ int main(int argc, char *argv[])
 	len1 = _strlen(s1);
 	len2 = _strlen(s2);
 	len = len1 + len2 + 1;
-	result = malloc(sizeof(int) * len);
+	result = malloc(sizeof(long int) * len);
 	if (!result)
 		return (1);
 	for (i = 0; i <= len1 + len2; i++)
